refactor(UserRepositoryTest): 사용자의 id로 북마크한 공지를 조회하는 메서드를 NoticeRepository로 이동한다
