+++ src/main/java/com/kustacks/kuring/user/domain/UserQueryRepositoryImpl.java
@@ -2,15 +2,12 @@ package com.kustacks.kuring.user.domain;
 
 import com.kustacks.kuring.admin.common.dto.FeedbackDto;
 import com.kustacks.kuring.admin.common.dto.QFeedbackDto;
-import com.kustacks.kuring.user.common.dto.BookmarkDto;
-import com.kustacks.kuring.user.common.dto.QBookmarkDto;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 
 import java.util.List;
 
-import static com.kustacks.kuring.notice.domain.QNotice.notice;
 import static com.kustacks.kuring.user.domain.QFeedback.feedback;
 
 @RequiredArgsConstructor
@@ -25,22 +22,6 @@ public class UserQueryRepositoryImpl implements UserQueryRepository {
                 .orderBy(feedback.createdAt.desc())
                 .offset(pageable.getOffset())
                 .limit(pageable.getPageSize())
-                .fetch();
-    }
-
-    @Override
-    public List<BookmarkDto> findAllBookmarks(List<String> ids) {
-        return queryFactory.select(
-                    new QBookmarkDto(
-                            notice.articleId,
-                            notice.postedDate,
-                            notice.subject,
-                            notice.categoryName.stringValue(),
-                            notice.url.value
-                    )
-                ).from(notice)
-                .where(notice.articleId.in(ids))
-                .orderBy(notice.postedDate.desc())
                 .fetch();
     }
 }
